+++ zappend/config/config.py
@@ -37,11 +37,11 @@ class Config:
         temp_storage_options = config_dict.get("temp_storage_options")
         self._temp_dir = FileObj(temp_dir_uri, storage_options=temp_storage_options)
 
-        # local import to avoid recursion
-        from ..slice.factory import to_slice_source_type
+        # avoid cyclic import
+        from ..slice.callable import to_slice_callable
 
         slice_source = config_dict.get("slice_source")
-        self._slice_source = to_slice_source_type(slice_source)
+        self._slice_source = to_slice_callable(slice_source)
 
     @property
     def zarr_version(self) -> int:
@@ -114,20 +114,24 @@ class Config:
 
     @property
     def slice_engine(self) -> str | None:
-        """The configured slice engine to be used if a slice object is not a Zarr.
+        """The configured slice engine to be used if a slice path or URI does not
+        point to a dataset in Zarr format.
         If defined, it will be passed to the `xarray.open_dataset()` function.
         """
         return self._config.get("slice_engine")
 
     @property
-    def slice_source(self) -> Callable | None:
-        """The configured slice source, if any."""
+    def slice_source(self) -> Callable[[...], Any] | None:
+        """The configured slice source type. If given, it must be
+        a callable that returns a value of type `SliceItem` or a class that is
+        derived from `SliceSource` abstract base class.
+        """
         return self._slice_source
 
     @property
     def slice_storage_options(self) -> dict[str, Any] | None:
         """The configured slice storage options to be used
-        if a slice object is a Zarr.
+        if a slice item is a URI.
         """
         return self._config.get("slice_storage_options")